+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## v1.0.2 [2018-07-04]
+
+  * Supports new API version 2018-04-10
+  * Smart stash support and stash description
+  * Add commit messages to metadata updates
+  * Add job state group `all` to list all jobs
+  * Extended VFX media families
+  * Extended list of media formats
+
 ## v1.0.1 [2018-04-11]
 
   * Supports new API version 2018-04-10